feat: implement lazy loading, add demo video in Home
implement lazy loading on components and images to improve initial load time
extract all loader and action functions to new files in new loader and action folder
wrap each page with its own <Suspense> to be able to implement their own loading page in the future
wrap each image with its own placeholder to preserve the space before loading is finished
remove 'element' property in routes config since it's unnecessary and would interfere lazy loading
previewadd  video and image in Home page to demonstrate the main feature of the website
enhance styling in DeleteModal, EditEntries,EditModal,InfoUpload,Modal
