+++ frontend/src/components/contest/ContestCompetePage..jsx
@@ -58,13 +58,3 @@ const ContestCompetePage = () => {
 };
 
 export default ContestCompetePage;
-
-export async function action({ request, params }) {
-  const entryId = await request.text();
-  const { contestId } = params;
-  await fetch(`http://localhost:8080/contest/${contestId}/entries/${entryId}`, {
-    method: "PATCH",
-  });
-
-  return null;
-}
