+++ frontend/src/loaders/loginLoader.js
@@ -1,0 +1,6 @@
+export function loginLoader({ request }) {
+  const searchParams = new URL(request.url).searchParams;
+  const prev = searchParams.get("from") || null;
+
+  return prev;
+}
