+++ frontend/src/loaders/rootLoader.js
@@ -1,5 +1,4 @@
-import { redirect } from "react-router-dom";
-import { store } from "../store/index";
+import { store } from "../store";
 import { userActions } from "../store/userSlice";
 
 export async function rootLoader() {
@@ -31,22 +30,4 @@ export async function rootLoader() {
     store.dispatch(userActions.login(user));
     return { isLogin: true, info: user };
   }
-}
-
-export async function rootAction() {
-  const response = await fetch("http://localhost:8080/auth/logout", {
-    method: "POST",
-    credentials: "include",
-  });
-
-  if (!response.ok) {
-    throw new Response(JSON.stringify({ message: "Logout failed, please try again later" }), {
-      status: 500,
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-  }
-  store.dispatch(userActions.logout());
-  return redirect("/");
 }
