+++ frontend/src/loaders/userLoader.js
@@ -1,0 +1,7 @@
+export const userLoader = async ({ params }) => {
+  const { userId } = params;
+
+  const response = await fetch(`http://localhost:8080/contest/user/${userId}`);
+
+  return response;
+};
