+++ frontend/src/pages/Edit.jsx
@@ -62,28 +62,3 @@ const Edit = () => {
 };
 
 export default Edit;
-
-export const loader = async ({ params }) => {
-  const { contestId } = params;
-  const response = await fetch(`http://localhost:8080/contest/${contestId}`);
-
-  return response;
-};
-
-export const action = async ({ request }) => {
-  const data = await request.json();
-  const { contestId } = data;
-  const options = { method: request.method, credentials: "include" };
-
-  if (request.method === "PATCH") {
-    options.headers = { "Content-Type": "application/json" };
-    options.body = JSON.stringify(data.contestContent);
-  }
-
-  const response = await fetch(`http://localhost:8080/contest/${contestId}`, options);
-  const { message } = await response.json();
-  if (!response.ok) {
-    return { successful: false, message };
-  }
-  return { successful: true, message };
-};
