+++ frontend/src/routes.jsx
@@ -1,9 +1,6 @@
-import BrowseContests from "./pages/BrowseContests";
-import CreateContest from "./pages/CreateContest";
-
 export const routes = [
-  { label: "Browse Contests", path: "/contests", element: <BrowseContests /> },
-  { label: "Create Contests", path: "/create", element: <CreateContest /> },
+  { label: "Browse Contests", path: "/contests" },
+  { label: "Create Contests", path: "/create" },
 ];
 
 export const dropDownRoutes = [{ label: "Home", path: "/" }, ...routes];
